backend/requirements.txt
# API
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1

# Processamento de dados
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2
pillow==10.2.0
plotly==5.19.0

# Desenvolvimento
python-dotenv==1.0.1
pytest==8.1.1
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0 